% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.ring.R
\name{check.ring}
\alias{check.ring}
\title{Checks the format of a RING database}
\usage{
check.ring(DNS1 = "Ring2017", DNS2 = "Ring2018", Lookup = "Lookup")
}
\arguments{
\item{DNS1}{RING Database used as standard}

\item{DNS2}{RING Database to check}

\item{Lookup}{Name of access driver for Lookup database. details.}
}
\description{
Checks that columns of database tables \strong{DNS2} are formatted as expected by comparing to a valid database which is used as a benchmark \strong{DNS1}. Further checks are implemented with respect to the definitions given in \strong{Lookup}, available \href{https://www.ifv-vogelwarte.de/fileadmin/resources/Beringerzentrale/Lookups.mde}{here}. See '\code{\link{export.ring}} for hints on the setup. Further guidance is provided in the documentation of \href{https://www.ifv-vogelwarte.de/fileadmin/resources/Beringerzentrale/RING_9.3-Programmdoku.pdf}{Ring 2.0}
}
