% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatR.R
\name{formatR}
\alias{formatR}
\title{format data}
\usage{
formatR(
  buzzard = "~/01-PhD/00-Raw/RData/buzzard_db.RData",
  kite = "~/../Dropbox/Fieldsaison_2020/Kite ringing list 2020-10-28 MO.xlsx",
  owls = "~/../Dropbox/Fieldsaison_2020/Eagle Owl ringing list 2020-08-28 MO.xlsx",
 
    others = "~/../Dropbox/Fieldsaison_2020/Goshawk and others ringing list 2020-08-28 MO.xlsx",
  year = 2020,
  exclude = 3419178
)
}
\arguments{
\item{buzzard}{path to buzzard_db.RData}

\item{kite}{path to Kite ringing list.xlsx}

\item{owls}{path to Eagle owl ringing list}

\item{others}{path to Goshawk and others ringing list.xlsx}

\item{year}{vector of years to export}

\item{exclude}{ring numbers to exclude from export (ie individuals without Vogelwartenring)}
}
\description{
format data
}
