% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place.info.R
\name{place.suggest}
\alias{place.suggest}
\title{Get place info for a coordinate}
\usage{
place.suggest(
  lat = NULL,
  lon = NULL,
  root = "C:/OSGeo4W64",
  Shape1 = NULL,
  Shape2 = NULL
)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{root}{location of RQGIS3}

\item{Shape2}{path to a ESRI shape file (.shp) with TK25. By default T25qdr.shp supplied as extdata with the package}

\item{GADM}{path to GADM ESRI shape file (.shp). By default GADM-NRW-NDS.shp, supplied as extdata with the package}
}
\description{
Executes the QGIS algorithm \code{qgis:joinattributesbylocation} via \code{\link[RQGIS3]{run_qgis}} to obtain place information from two vector layers. For example, administrative levels from \href{https://gadm.org/}{gadm.org}) and corresponding TK25 numbers (e.g. available from \href{https://www.ornitho.de/index.php?m_id=1118&item=15}{here}).
}
