% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Push into a ring}
\usage{
push(x, data, ...)
}
\arguments{
\item{x}{A ring buffer of some sort}

\item{data}{Data to push into the ring}

\item{...}{Additional arguments passed through to methods}
}
\description{
Push elements onto a ring object
}

