% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.lookup.R
\name{read.lookup}
\alias{read.lookup}
\title{Open all lookup tables as a tab}
\usage{
read.lookup(Lookup = NULL)
}
\arguments{
\item{Lookup}{Name of lookups.mde driver}
}
\value{
list of data frame
}
\description{
Open all lookup tables as a tab
}
\details{
See \link{export.ring} and \link{check.ring}
}
