% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_buffer_bytes.R
\name{ring_buffer_bytes}
\alias{ring_buffer_bytes}
\title{Byte array based ring buffer}
\usage{
ring_buffer_bytes(size, stride = 1L)
}
\arguments{
\item{size}{Size of the buffer, each entry of which will be
\code{stride} bytes long.}

\item{stride}{Number of bytes per buffer entry.  Defaults to 1
byte.}
}
\description{
Construct a ring buffer where the buffer holds a stream of bytes.
Optionally the buffer can be "strided" so that the bytes naturally
fall into chunks of exactly the same size.  It is implemented in C
and will be fast, with the limitation that any data transfer
always involves copies.
}

