% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_buffer_bytes_typed.R
\name{ring_buffer_bytes_typed}
\alias{ring_buffer_bytes_typed}
\title{Typed bytes buffer}
\usage{
ring_buffer_bytes_typed(size, what, len = NULL)
}
\arguments{
\item{size}{The maximum number of elements the buffer can hold.
Each element will be multiple bytes long.}

\item{what}{Either a vector on the style of \code{vapply} (e.g.,
\code{integer(4)} to indicate that each element of the buffer is
a 4-element integer, or the \code{name} of a storage mode if
\code{len} is also provided.}

\item{len}{If given, then the length of the storage.  If it is
given, then if \code{length(what)} is zero, the storage mode of
\code{what} is used as the type.  Otherwise \code{what} is
interpreted as the \emph{name} of the storage mode (one of
"logical", "integer", "double" or "complex".}
}
\description{
Create a ring buffer, backed by a byte array, where each element
corresponds to a vector of one of R's atomic numeric types
(logical, integer, double, and complex).
}
\details{
Note that a logical ring buffer and an integer ring buffer take
the same number of bytes because a logical vector is stored as an
integer (4 bytes per element) to deal with missing values.
}
\author{
Rich FitzJohn
}

