% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_buffer_env.R
\name{ring_buffer_env}
\alias{ring_buffer_env}
\title{Environment-based ring buffer}
\usage{
ring_buffer_env(size, prevent_overflow = FALSE)
}
\arguments{
\item{size}{The number of entries the buffer can contain.}

\item{prevent_overflow}{Logial indicating if buffer overflow is
not allowed.  If \code{FALSE} (the default) then the buffer will
overflow silently (that is, the oldest data will be
overwritten).  If \code{TRUE}, then on overflow an error will be
raised.}
}
\description{
An environment based ring buffer.  In contrast with the bytes
buffer, this one is truely circular, implemented as a pair of
doubly linked lists.
}
\examples{
buf <- ring_buffer_env(10)
buf$push(1:10)
buf$take(3)
buf$push(11:15)
buf$take(2)
}
\author{
Rich FitzJohn
}

