% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.shape.R
\name{unlink.shp}
\alias{unlink.shp}
\title{Remove temporary shape files}
\usage{
unlink.shp(file = NULL, fileEnding = ".shp")
}
\arguments{
\item{file}{file name}

\item{fileEnding}{default .shp}
}
\description{
Remove temporary shape files
}
