% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_buffer_env.R
\name{ring_buffer_env}
\alias{ring_buffer_env}
\title{Environment-based ring buffer}
\usage{
ring_buffer_env(size)
}
\arguments{
\item{size}{The number of entries the buffer can contain.}
}
\description{
An environment based ring buffer.  In contrast with the bytes
buffer, this one is truely circular, implemented as a pair of
doubly linked lists.
}
\examples{
buf <- ring_buffer_env(10)
buf$push(1:10)
buf$take(3)
buf$push(11:15)
buf$take(2)
}
\author{
Rich FitzJohn
}

