% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_vector.R
\name{ring_vector}
\alias{ring_vector}
\title{Ring vector}
\usage{
ring_vector(length_max, type, environment = TRUE)
}
\arguments{
\item{length_max}{The maximum number of elements}

\item{type}{The type of storage.  Can be "logical", "integer",
"double", or "complex"}

\item{environment}{Logical indicating if we should use an
environment buffer (\code{\link{ring_buffer_env}}) or
a bytes buffer (\code{\link{ring_buffer_bytes}}).}
}
\description{
Simulate an atomic vector with a ring buffer.  This exists mostly
as an example of use of a ring buffer designed to work with R
functions that do not know (or care) that the object is
implemented with a ring buffer behind the scenes.  Elements will
be added at the end of the vector and taken from the beginning.
}

